Initial commit
This is the projects I have done in the class Pattern Recognition and
Machine Learning. The first project is to use PCA to model human’s face
and use Fisher Linear Discrimination to classify female/male. The
second project is to use Adaboost and Realboost to train classifier for
human face detection. More information is available here
http://www.stat.ucla.edu/~sczhu/Courses/UCLA/Stat_231/Stat_231.html